Cleanup and A* optimal path bug fix
Fixed bug causing A* to sometimes not provide the optimal path. The problem related to how the __lt__, __gt__, and __eq__ methods of the node class were set up. Instead of solely comparing the sum of the G and H attributes, there are cases when the H attribute should take precedence.

I also cleaned up the program, adding comments where necessary and removing dead code.